actually fixes [SelfHostedPro#328]
My previous PR left errors with setting usernames that I missed, but now it
works as intended.  I did not get the user password don't change on an
empty string because I think it would require a change in how the user
is handled in the backend.  I also cleaned up a print function here that
was printing an ugly object reference on a name change, so now it shows old/new
usernames if they are changing.